requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
cvzone>=1.5.6
ultralytics>=8.1.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
tensorflow>=2.13.0
protobuf>=3.20.0